factor out analytics method of log module into own module
