+++ lib/app.js
@@ -11,6 +11,7 @@ var rimraf = require('rimraf');
 var Progger = require('progger');
 var util = require(path.resolve(__dirname, './', 'util'));
 var msg = require(path.resolve(__dirname, './', 'msg'));
+var analytics = require(path.resolve(__dirname, './', 'analytics'));
 var p = new Progger({ speed: 100, token: '.', color: 'blue' });
 
 module.exports = function (args, opts) {
@@ -354,7 +355,7 @@ module.exports = function (args, opts) {
     // log the process analytics
     if (opts.analytics && showAnalytics) {
       log.appEnd = microtime.now();
-      return msg.analytics(log);
+      return analytics(log);
     }
   }).done();
 };
